Using variables in custom/separate Widgets
