+++ first_app/lib/main.dart
@@ -1,5 +1,6 @@
 // Way to import a package in Dart
 import 'package:flutter/material.dart';
+import 'package:first_app/gradient_container.dart';
 
 // Main function. Entrypoint of Dart App
 void main() {
@@ -19,34 +20,4 @@ void main() {
       ),
     ),
   );
-}
-
-// GradientContainer. Widget for separate Gradient background of screen
-//    * StatelessWidget. Flutter class for built a widget
-class GradientContainer extends StatelessWidget {
-  const GradientContainer({super.key});
-
-  // build method automatically receive by Flutter a values of context
-  // context. Object that containts useful information
-  @override
-  Widget build(context) {
-    return Container(
-      decoration: const BoxDecoration(
-        gradient: LinearGradient(
-          colors: [
-            Color.fromARGB(255, 0, 77, 77),
-            Color.fromARGB(255, 0, 89, 116),
-          ],
-          begin: Alignment.topLeft,
-          end: Alignment.bottomRight,
-        ),
-      ),
-      child: const Center(
-        child: Text(
-          'Hallo Wald!',
-          style: TextStyle(fontSize: 28, color: Colors.white),
-        ),
-      ),
-    );
-  }
 }
