+++ first_app/lib/styled_text.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+class StyledText extends StatelessWidget {
+  const StyledText({super.key});
+
+  @override
+  Widget build(context) {
+    return const Text(
+      'Hey You 🌟!',
+      style: TextStyle(fontSize: 28, color: Colors.white),
+    );
+  }
+}
